+++ src/squadMay21/stats/Variance.java
@@ -3,7 +3,7 @@ package squadMay21.stats;
 public class Variance {
 
 	
-	public static int getVariance(int[] arr) {
+	public static double getVariance(int[] arr) {
 		double avg = Mean.getMean(arr);
 		int v = 0;
 		for(int n : arr) {
